requirements.txt
django>=4.2,<5.0
psycopg2-binary>=2.9
pgvector>=0.2.5
groq>=0.9.0
django-cors-headers>=4.3.0
python-dotenv>=1.0.0
python-decouple>=3.8
gunicorn>=21.0.0
sentence-transformers>=3.0.0
torch>=2.0.0
